refactor redux store
