+++ client/src/app/redux.tsx
@@ -1,102 +1,24 @@
-import { useRef } from "react";
-import { combineReducers, configureStore } from "@reduxjs/toolkit";
-import {
-  TypedUseSelectorHook,
-  useDispatch,
-  useSelector,
-  Provider,
-} from "react-redux";
-import globalReducer from "@/state";
-// import { api } from "@/state/api";
-import { setupListeners } from "@reduxjs/toolkit/query";
+'use client';
 
-import {
-  persistStore,
-  persistReducer,
-  FLUSH,
-  REHYDRATE,
-  PAUSE,
-  PERSIST,
-  PURGE,
-  REGISTER,
-} from "redux-persist";
+import { Provider } from "react-redux";
+import { store } from "@/utils/store";
+import { persistStore } from "redux-persist";
 import { PersistGate } from "redux-persist/integration/react";
-import createWebStorage from "redux-persist/lib/storage/createWebStorage";
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
+import type { RootState, AppDispatch } from "@/utils/store";
 
-/* REDUX PERSISTENCE */
-const createNoopStorage = () => {
-  return {
-    getItem() {
-      return Promise.resolve(null);
-    },
-    setItem(_key: string, value: string) {
-      return Promise.resolve(value);
-    },
-    removeItem() {
-      return Promise.resolve();
-    },
-  };
-};
-
-const storage =
-  typeof window === "undefined"
-    ? createNoopStorage()
-    : createWebStorage("local");
-
-const persistConfig = {
-  key: "root",
-  storage,
-  whitelist: ["global"],
-};
-const rootReducer = combineReducers({
-  global: globalReducer,
-  // [api.reducerPath]: api.reducer,
-});
-const persistedReducer = persistReducer(persistConfig, rootReducer);
-
-/* REDUX STORE */
-/**
- * Creates and configures the Redux store with persisted reducer and middleware.
- * @returns {EnhancedStore} The configured Redux store.
-*/
-export const makeStore = () => {
-  return configureStore({
-    // Use the persisted reducer to maintain state across sessions
-    reducer: persistedReducer,
-    // Configure middleware with serializable checks and API middleware
-    middleware: (getDefaultMiddleware) =>
-      getDefaultMiddleware({
-        serializableCheck: {
-          // Ignore these actions for the serializable check
-          ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER],
-        },
-      })
-      // .concat(api.middleware), // Add API middleware for handling queries
-  });
-};
-
-/* REDUX TYPES */
-export type AppStore = ReturnType<typeof makeStore>;
-export type RootState = ReturnType<AppStore["getState"]>;
-export type AppDispatch = AppStore["dispatch"];
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
 
-/* PROVIDER */
 export default function StoreProvider({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const storeRef = useRef<AppStore>();
-  if (!storeRef.current) {
-    storeRef.current = makeStore();
-    setupListeners(storeRef.current.dispatch);
-  }
-  const persistor = persistStore(storeRef.current);
+  const persistor = persistStore(store);
 
   return (
-    <Provider store={storeRef.current}>
+    <Provider store={store}>
       <PersistGate loading={null} persistor={persistor}>
         {children}
       </PersistGate>
